requirements.txt
django
djangorestframework
markdown
django-filter
django-webpack-loader
mysqlclient
djangorestframework-simplejwt
simplejwt
requests
six
